+++ spec/integrations_helper.rb
@@ -21,7 +21,10 @@ def setup_karafka
     # Use some decent defaults
     caller_id = [caller_locations(1..1).first.path.split('/').last, SecureRandom.uuid].join('-')
 
-    config.kafka = { 'bootstrap.servers' => '127.0.0.1:9092' }
+    config.kafka = {
+      'bootstrap.servers' => '127.0.0.1:9092',
+      'statistics.interval.ms' => 100
+    }
     config.client_id = caller_id
     config.pause_timeout = 1
     config.pause_max_timeout = 1
@@ -29,12 +32,15 @@ def setup_karafka
     config.max_wait_time = 500
     config.shutdown_timeout = 30_000
     config.producer = ::WaterDrop::Producer.new do |producer_config|
-      producer_config.kafka = config.kafka
+      # In some cases WaterDrop updates the config and we don't want our consumer config to be
+      # polluted by those updates, that's why we copy
+      producer_config.kafka = config.kafka.dup
       producer_config.logger = config.logger
       # We need to wait a lot sometimes because we create a lot of new topics and this can take
       # time
       producer_config.max_wait_timeout = 120 # 2 minutes
     end
+
 
     # Allows to overwrite any option we're interested in
     yield(config) if block_given?
@@ -42,8 +48,13 @@ def setup_karafka
 
   Karafka.logger.level = 'debug'
 
+  # We turn on all the instrumentation just to make sure it works also in the integration specs
   Karafka.monitor.subscribe(Karafka::Instrumentation::StdoutListener.new)
   Karafka.monitor.subscribe(Karafka::Instrumentation::ProctitleListener.new)
+
+  # We turn on also WaterDrop instrumentation the same way and for the same reasons as above
+  listener = ::WaterDrop::Instrumentation::StdoutListener.new(Karafka.logger)
+  Karafka.producer.monitor.subscribe(listener)
 end
 
 # Waits until block yields true