+++ spec/spec_helper.rb
@@ -4,8 +4,6 @@ ENV['KARAFKA_ENV'] = 'test'
 $LOAD_PATH.unshift(File.dirname(__FILE__))
 $LOAD_PATH.unshift(File.join(File.dirname(__FILE__), '..', 'lib'))
 
-# @note HashWithIndifferentAccess is just for testing the optional integration,
-# it is not used by default in the framework
 %w[
   byebug
   factory_bot