+++ CHANGELOG.md
@@ -36,6 +36,8 @@
 - `WaterDrop` `2.0` support
 - Simple routing style (`0.5`) now builds a single consumer group instead of one per topic
 - Example apps were updated
+- Hook for underlying statistics emitted from librdkafka have been added.
+- Hook for underlying async errors emitted from  librdkafka have been added.
 
 ## 1.4.7 (2021-09-04)
 - Update ruby-kafka to `1.4.0`